+++ src/main/java/domain/Deck.java
@@ -1,73 +1,41 @@
 package domain;
 
 import common.ErrorMessage;
-import java.util.ArrayList;
+import java.util.Deque;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 public class Deck {
-    private static final int BUST_CRITERIA = 21;
-    private final List<Card> cards;
+    private final Deque<Card> cards;
 
-    private Deck(List<Card> cards) {
+    private Deck(Deque<Card> cards) {
         this.cards = cards;
     }
 
     public static Deck createDeck(CardCreationStrategy strategy) {
-        List<Card> cards = strategy.create();
-        return new Deck(cards);
+        return new Deck(strategy.create());
     }
 
-    public static Deck createParticipantDeck(Deck totaldeck) {
-        List<Card> cards = new ArrayList<>(
-                List.of(
-                        totaldeck.drawCard(),
-                        totaldeck.drawCard()
-                )
-        );
-        return new Deck(cards);
-    }
-
-    public List<Card> getCards() {
-        return cards;
-    }
+    // 메시지 : 카드를 만들어줘.
 
     public Card drawCard() {
-        if (cards.isEmpty()) {
+        try {
+            return cards.pop();
+        } catch (NoSuchElementException e) {
             throw new IllegalArgumentException(ErrorMessage.DRAW_CARD_OUT_OF_RANGE.getMessage());
         }
-
-        return cards.removeFirst();
     }
 
-    public boolean isBust() {
-        return calculateCardScoreSum() > BUST_CRITERIA;
-    }
-
-    public int calculateCardScoreSum() {
-        int sumExceptAce = calculateCardScoreSumExceptAce();
-        int sumAce = new AceScoreDiscriminator().calculateAceCardsSum(cards, sumExceptAce);
-
-        return sumAce + sumExceptAce;
+    public List<Card> drawTwoCards() {
+        try {
+            return List.of(cards.pop(), cards.pop());
+        } catch (NoSuchElementException e) {
+            throw new IllegalArgumentException(ErrorMessage.DRAW_CARD_OUT_OF_RANGE.getMessage());
+        }
     }
 
     public int addCard(Card card) {
         this.cards.add(card);
         return cards.size();
-    }
-
-    private int calculateCardScoreSumExceptAce() {
-        int sum = 0;
-        for (Card card : cards) {
-            sum = addCardScoreExceptAce(card, sum);
-        }
-
-        return sum;
-    }
-
-    private int addCardScoreExceptAce(Card card, int sum) {
-        if (!card.isAce()) {
-            sum += card.getCardContents().getScore();
-        }
-        return sum;
     }
 }
