+++ src/main/java/domain/Hand.java
@@ -1,0 +1,46 @@
+package domain;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Hand {
+    private static final int BUST_CRITERIA = 21;
+    private final List<Card> cards;
+
+    private Hand(List<Card> cards) {
+        this.cards = new ArrayList<>(cards);
+    }
+
+    public static Hand of(Card card1, Card card2) {
+        return new Hand(List.of(card1, card2));
+    }
+
+    //사용자 카드 보관함을 생성해줘!!
+
+    public void addCard(Card card) {
+        this.cards.add(card);
+    }
+
+    public int calculateCardScoreSum() {
+        int sumExceptAce = calculateCardScoreSumExceptAce();
+        int sumAce = AceScoreDiscriminator.calculateAceCardsSum(cards, sumExceptAce);
+
+        return sumAce + sumExceptAce;
+    }
+
+    private int calculateCardScoreSumExceptAce() {
+        int sum = 0;
+        for (Card card : cards) {
+            sum = addCardScoreExceptAce(card, sum);
+        }
+
+        return sum;
+    }
+
+    private int addCardScoreExceptAce(Card card, int sum) {
+        if (!card.isAce()) {
+            sum += card.getCardContents().getScore();
+        }
+        return sum;
+    }
+}
